switch to python-memcached
