+++ bin/polaris-health-control
@@ -6,7 +6,7 @@ import socket
 
 import daemon
 
-from  polaris_health import config, runtime
+from  polaris_health import config, guardian
 
 
 # control socket recv() buffer size
@@ -35,7 +35,7 @@ class PolarisHealthControl:
 
     def __init__(self):
         # load polaris_health.config
-        runtime.Runtime.load_configuration()
+        guardian.Guardian.load_configuration()
 
     def ping(self):
         self._send_control_sock_cmd('ping') == 'pong'
@@ -48,11 +48,11 @@ class PolarisHealthControl:
             raise Exception('Polaris health is already running')
 
         if debug:
-            runtime.Runtime().start(debug=True)
+            guardian.Guardian().start(debug=True)
         else:
             # turn into a daemon, NO CODE WILL EXECUTE AFTER THIS
             with daemon.DaemonContext():
-                runtime.Runtime().start()
+                guardian.Guardian().start()
 
     def start_debug(self):
         self.start(debug=True)
